service/requirements.txt
boto3==1.20.23
celery==5.2.1
dj-database-url==0.5.0
django-ace==1.0.11
django-admin-autocomplete-filter==0.7.1
django-admin-interface==0.18.3
django-apptemplates==1.5
django-clearcache==1.2.1
django-colorfield==0.5.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-phonenumber-field==6.0.0
django-storages==1.12.3
django==4.0.0
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
phonenumbers==8.12.39
psycopg2-binary==2.9.2
PyJWT==2.3.0
redis==4.0.2
requests==2.26.0
sentry-sdk==1.5.0
twilio==7.3.2
whitenoise==5.3.0
